Break out upgrade logic
